==> scripts/0_utils.R <==
# Utility functions for PharmaPayments

library(dplyr)
library(stringr)

clean_specialties <- function(data) {
  keyword_to_specialty <- c(
"ADDICTION" = "ADDICTION MEDICINE",
"ANAESTHES" = "ANAESTHESIA",
"CARDIOL" = "CARDIOLOGY",
"THORACIC SURGERY" = "CARDIO-THORACIC SURGERY",
"CLINICAL GENETIC" = "CLINICAL GENETICS",
"CLINICAL PHARM" = "CLINICAL PHARMACOLOGY",
"COLORECTAL SURG" = "COLORECTAL SURGERY",
"DERMATOLOG" = "DERMATOLOGY",
"RADIOLOGY" = "DIAGNOSTIC RADIOLOGY",
"DOCTOR-IN-TRAINING" = "DOCTOR-IN-TRAINING",
"EMERGENCY" = "EMERGENCY MEDICINE",
"ENDOCRIN" = "ENDOCRINOLOGY",
"GASTRO" = "GASTROENTEROLOGY AND HEPATOLOGY",
"GENERAL MEDICINE" = "GENERAL MEDICINE",
"GP" = "GENERAL PRACTICE",
"GENERAL SURG" = "GENERAL SURGERY",
"GERIATRIC" = "GERIATRIC MEDICINE",
"HAEMATOLOG" = "HAEMATOLOGY_ONCOLOGY",
"IMMUNOLOG" = "IMMUNOLOGY AND ALLERGY",
"INFECTIOUS" = "INFECTIOUS DISEASES",
"INTENSIVE CARE" = "INTENSIVE CARE MEDICINE",
"MEDICAL ADMINISTRATION" = "MEDICAL ADMINISTRATION",
"MICROBIOLOG" = "MICROBIOLOGY",
"NEPHROLO" = "NEPHROLOGY",
"NEUROLOGY" = "NEUROLOGY",
"NEUROSURG" = "NEUROSURGERY",
"NUCLEAR" = "NUCLEAR MEDICINE",
"ONCOLOG" = "HAEMATOLOGY_ONCOLOGY",
"OCCUPATIONAL" = "OCCUPATIONAL MEDICINE",
"OBSTETRIC" = "OBSTETRICS AND GYNAECOLOGY",
"OPHTHALM" = "OPHTHALMOLOGY",
"ORTHOPAEDIC" = "ORTHOPAEDIC SURGERY",
"MAXILLOFACIAL" = "ORAL AND MAXILLOFACIAL SURGERY",
"OTOLARYNGOL" = "OTOLARYNGOLOGY - HEAD AND NECK SURGERY",
"PAEDIATRIC ENDOCRINOL" = "PAEDIATRIC ENDOCRINOLOGY",
"PAEDIATRIC GASTRO" = "PAEDIATRIC GASTROENTEROLOGY AND HEPATOLOGY",
"PAEDIATRIC HAEMAT" = "PAEDIATRIC HAEMATOLOGY_ONCOLOGY",
"PAEDIATRIC IMMUNOLO" = "PAEDIATRIC IMMUNOLOGY AND ALLERGY",
"PAEDIATRIC INFECTIOUS" = "PAEDIATRIC INFECTIOUS DISEASES",
"PAEDIATRIC MEDICAL ONCOLOGY" = "PAEDIATRIC MEDICAL ONCOLOGY",
"PAEDIATRIC NEUROL" = "PAEDIATRIC NEUROLOGY",
"PAEDIATRIC RESPIRATO" = "PAEDIATRIC RESPIRATORY AND SLEEP MEDICINE",
"PAEDIATRIC RHEUMAT" = "PAEDIATRIC RHEUMATOLOGY",
"PAEDIATRIC SURGERY" = "PAEDIATRIC SURGERY",
"PAEDIATRICS AND CHILD HEALTH" = "PAEDIATRICS AND CHILD HEALTH",
"PAEDRIATICS" = "PAEDIATRICS",
"PAIN MEDICINE" = "PAIN MEDICINE",
"PALLIATIVE MEDICINE" = "PALLIATIVE MEDICINE",
"PATHOLOGY" = "PATHOLOGY",
"PHYSICIAN" = "PHYSICIAN",
"PLASTIC" = "PLASTIC SURGERY",
"PSYCHIAT" = "PSYCHIATRY",
"PUBLIC HEALTH" = "PUBLIC HEALTH MEDICINE",
"RADIATION ONCOL" = "RADIATION ONCOLOGY",
"REHABILITATION" = "REHABILITATION MEDICINE",
"RESPIRATORY" = "RESPIRATORY AND SLEEP MEDICINE",
"RHEUMATOLO" = "RHEUMATOLOGY",
"SEXUAL HEALTH" = "SEXUAL HEALTH MEDICINE",
"SPORT AND EXERCISE" = "SPORT AND EXERCISE MEDICINE",
"SURGE" = "SURGERY",
"UROLOG" = "UROLOGY",
"VASCULAR" = "VASCULAR SURGERY",
"GP, GENERAL PRACT" = "GENERAL PRACTICE",
"GENERAL PRACT" = "GENERAL PRACTICE",
"PAIN MEDICINE, PHYSICIAN" = "PAIN MEDICINE",
"GASTRO, HEPATO" = "GASTROENTEROLOGY AND HEPATOLOGY",
"ORTHOPAEDIC, SURGE" = "ORTHOPAEDIC SURGERY",
"COLORECTAL SURG, SURGE" = "COLORECTAL SURGERY",
"CARDIOL, VASCULAR" = "CARDIOLOGY",
"ONCOLOG" = "HAEMATOLOGY_ONCOLOGY",
"PHYSICIAN" = "GENERAL MEDICINE",
"GENERAL SURG, SURGE" = "GENERAL SURGERY",
"OTOLARYNGOL" = "OTOLARYNGOLOGY - HEAD AND NECK SURGERY",
"CARDIOL, PHYSICIAN" = "CARDIOLOGY",
"NEUROLOGY, UROLOG" = "NEUROLOGY",
"IMMUNOLOG, ALLERG" = "IMMUNOLOGY AND ALLERGY",
"PHYSICIAN, VASCULAR" = "GENERAL MEDICINE",
"THORACIC SURGERY, SURGE" = "THORACIC SURGERY",
"ENDOCRIN, PHYSICIAN" = "ENDOCRINOLOGY",
"DERMATOLOG, SURGE" = "DERMATOLOGY",
"RADIOLOGY, NUCLEAR" = "RADIOLOGY",
"HAEMATOLOG, ONCOLOG" = "HAEMATOLOGY_ONCOLOGY",
"OBSTETRIC, GYNAECOL" = "OBSTETRICS AND GYNAECOLOGY",
"PHYSICIAN, RESPIRATORY" = "RESPIRATORY AND SLEEP MEDICINE",
"GYNAECOL" = "OBSTETRICS AND GYNAECOLOGY",
"CARDIOL, GASTRO" = "CARDIOLOGY",
"EYE" = "OPHTHALMOLOGY",
"DERMATOLOG, GASTRO" = "DERMATOLOGY",
"DERMATOLOG, PLASTIC, SURGE" = "DERMATOLOGY",
"RENAL" = "NEPHROLOGY",
"RADIATION ONCOL, ONCOLOG" = "RADIATION ONCOLOGY",
"SURGE, EYE" = "SURGERY",
"EMERGENCY, PHYSICIAN" = "EMERGENCY MEDICINE",
"ENDOCRIN, PAEDIATRIC ENDOCRINOL" = "PAEDIATRIC ENDOCRINOLOGY",
"ENDOCRIN, PATHOLOGY" = "ENDOCRINOLOGY",
"ENDOCRIN, SURGE, RENAL" = "ENDOCRINOLOGY",
"ENDOCRIN, OBSTETRIC, PHYSICIAN" = "ENDOCRINOLOGY",
"ENDOCRIN, NUCLEAR, PHYSICIAN" = "ENDOCRINOLOGY",
"ENDOCRIN, OBSTETRIC" = "ENDOCRINOLOGY",
"ENDOCRIN, GERIATRIC" = "ENDOCRINOLOGY",
"ENDOCRIN, NEUROLOGY, UROLOG" = "NEUROLOGY",
"SURGE, VASCULAR" = "VASCULAR SURGERY",
"ENDOCRIN, EYE" = "ENDOCRINOLOGY",
"CARDIOL, ENDOCRIN" = "CARDIOLOGY",
"ENDOCRIN, NUCLEAR" = "ENDOCRINOLOGY",
"HEPATO" = "GASTROENTEROLOGY AND HEPATOLOGY",
"GASTRO, SURGE" = "GASTROENTEROLOGY AND HEPATOLOGY",
"GASTRO, PHYSICIAN" = "GASTROENTEROLOGY AND HEPATOLOGY",
"CARDIOL, RESPIRATORY" = "CARDIOLOGY",
"GASTRO, GENERAL MEDICINE" = "GASTROENTEROLOGY AND HEPATOLOGY",
"NEUROLOGY, PHYSICIAN, UROLOG" = "NEUROLOGY",
"GP, EYE" = "GP / GENERAL PRACTICE",
"GP, SURGE" = "GP / GENERAL PRACTICE",
"GP, OCCUPATIONAL" = "GP / GENERAL PRACTICE",
"PHYSICIAN, SEXUAL HEALTH" = "SEXUAL HEALTH MEDICINE",
"GP, SEXUAL HEALTH" = "GP / GENERAL PRACTICE",
"GP, PHYSICIAN" = "GP / GENERAL PRACTICE",
"GP, OBSTETRIC" = "GP / GENERAL PRACTICE",
"PLASTIC, SURGE" = "PLASTIC SURGERY",
"ENDOCRIN, RESPIRATORY" = "ENDOCRINOLOGY",
"DERMATOLOG, GP" = "DERMATOLOGY",
"ALLERG" = "IMMUNOLOGY AND ALLERGY",
"GASTRO, GENERAL SURG, SURGE" = "GASTROENTEROLOGY AND HEPATOLOGY",
"ENDOCRIN, SURGE" = "ENDOCRINOLOGY",
"GENERAL SURG, SURGE, ONCOLOG" = "GENERAL SURGERY",
"SURGE, ONCOLOG" = "SURGERY",
"GERIATRIC, REHABILITATION" = "GERIATRIC MEDICINE",
"HAEMATOLOG, PAEDIATRIC HAEMAT" = "PAEDIATRIC HAEMATOLOGY_ONCOLOGY",
"PHYSICIAN, ONCOLOG" = "HAEMATOLOGY_ONCOLOGY",
"HAEMATOLOG, PATHOLOGY" = "HAEMATOLOGY_ONCOLOGY",
"OBSTETRIC, ONCOLOG, GYNAECOL" = "OBSTETRICS AND GYNAECOLOGY",
"MAXILLOFACIAL, SURGE" = "ORAL AND MAXILLOFACIAL SURGERY",
"IMMUNOLOG, INFECTIOUS" = "IMMUNOLOGY AND ALLERGY",
"INFECTIOUS, PHYSICIAN" = "INFECTIOUS DISEASES",
"ENDOCRIN, RENAL" = "ENDOCRINOLOGY",
"NEPHROLO, RENAL" = "NEPHROLOGY",
"GENERAL MEDICINE, NEPHROLO, PHYSICIAN" = "GENERAL MEDICINE",
"NEUROSURG, SURGE" = "NEUROSURGERY",
"ENDOCRIN, NEPHROLO" = "ENDOCRINOLOGY",
"NEPHROLO, PHYSICIAN" = "NEPHROLOGY",
"REHABILITATION, UROLOG" = "REHABILITATION MEDICINE",
"NEUROLOGY, NEUROSURG, SURGE, UROLOG" = "NEUROLOGY",
"UROLOG, EYE" = "OPHTHALMOLOGY",
"OPHTHALM, SURGE, EYE" = "OPHTHALMOLOGY",
"NEUROLOGY, PSYCHIAT, UROLOG" = "PSYCHIATRY",
"NEUROSURG, SURGE, UROLOG" = "NEUROSURGERY",
"NUCLEAR, PHYSICIAN" = "NUCLEAR MEDICINE",
"OPHTHALM, EYE" = "OPHTHALMOLOGY",
"SURGE, UROLOG" = "SURGERY",
"MICROBIOLOG, PUBLIC HEALTH" = "PUBLIC HEALTH MEDICINE",
"SURGE, GYNAECOL" = "SURGERY",
"ONCOLOG, GYNAECOL" = "HAEMATOLOGY_ONCOLOGY",
"OBSTETRIC, SURGE, GYNAECOL" = "OBSTETRICS AND GYNAECOLOGY",
"OPHTHALM, SURGE" = "OPHTHALMOLOGY",
"PLASTIC, SURGE, EYE" = "PLASTIC SURGERY",
"PHYSICIAN, SURGE, EYE" = "PHYSICIAN",
"OPHTHALM, PLASTIC, SURGE" = "OPHTHALMOLOGY",
"GP, OPHTHALM" = "GP / GENERAL PRACTICE",
"NEUROLOGY, PAEDIATRIC NEUROL, UROLOG" = "NEUROLOGY",
"GASTRO, PAEDIATRIC GASTRO" = "GASTROENTEROLOGY AND HEPATOLOGY",
"PAEDIATRIC NEUROL, UROLOG" = "PAEDIATRIC NEUROLOGY",
"PAEDIATRIC RESPIRATO, RESPIRATORY" = "PAEDIATRIC RESPIRATORY AND SLEEP MEDICINE",
"PAEDIATRIC RHEUMAT, RHEUMATOLO" = "RHEUMATOLOGY",
"PHYSICIAN, REHABILITATION" = "REHABILITATION MEDICINE",
"GP, PAIN MEDICINE, PHYSICIAN" = "GP / GENERAL PRACTICE",
"GASTRO, PATHOLOGY" = "GASTROENTEROLOGY AND HEPATOLOGY",
"PSYCHIAT, EYE" = "PSYCHIATRY",
"INTENSIVE CARE, PHYSICIAN, RESPIRATORY" = "INTENSIVE CARE MEDICINE",
"GP, RESPIRATORY" = "GP / GENERAL PRACTICE",
"ORTHOPAEDIC, RHEUMATOLO" = "ORTHOPAEDIC SURGERY",
"GERIATRIC, RHEUMATOLO" = "GERIATRIC MEDICINE",
"PHYSICIAN, RHEUMATOLO" = "PHYSICIAN",
"OBSTETRIC, RHEUMATOLO" = "OBSTETRICS AND GYNAECOLOGY",
"PHYSICIAN, SPORT AND EXERCISE" = "SPORT AND EXERCISE MEDICINE"
  )

  data %>%
    mutate(
      matched_full_specialty = keyword_to_specialty[toupper(matched_keywords_c)],
      final_spec = str_to_upper(final_spec)
    ) %>%
    mutate(
      final_spec = case_when(
        final_spec == "SOCIAL WORKER" ~ "SOCIAL WORK",
        final_spec == "PHYSIOTHERAPY AND OCCUPATIONAL THERAPY" ~ "PHYSIOTHERAPY OR OCCUPATIONAL THERAPY",
        final_spec == "PHYSIOTHERAPIST OR OCCUPATIONAL THERAPIST" ~ "PHYSIOTHERAPY OR OCCUPATIONAL THERAPY",
        final_spec == "PHARMACIST" ~ "PHARMACY",
        final_spec == "MEDICAL ONCOLOGY" ~ "HAEMATOLOGY_ONCOLOGY",
        final_spec == "GP" ~ "GENERAL PRACTICE",
        final_spec == "GENERAL PRACTITIONER" ~ "GENERAL PRACTICE",
        final_spec == "NON-CLINICAL" ~ "SCIENTIST",
        final_spec == "RADIOLOGY" ~ "DIAGNOSTIC RADIOLOGY",
        final_spec == "SONOGRAPHER" ~ "RADIOGRAPHY",
        final_spec == "PSYCHOLOGIST" ~ "PSYCHOLOGY",
        TRUE ~ final_spec
      )
    )
}

standardize_service <- function(data) {
  service_mapping <- c(
    # Advisory Board Attendee
    "adboard or committee meeting" = "Advisory Board Attendee",
    "advisory board or committee meeting" = "Advisory Board Attendee",
    "advisory board/committee meeting" = "Advisory Board Attendee",
    "advisory board orcommitteemeeting" = "Advisory Board Attendee",
    "advisory board / committee member" = "Advisory Board Attendee",
    "advisory board committee member" = "Advisory Board Attendee",
    "advisory board member" = "Advisory Board Attendee",
    "advisory board" = "Advisory Board Attendee",
    "advisory board/ committee member overseas" = "Advisory Board Attendee",
    "advisory board/committee member" = "Advisory Board Attendee",
    "advisory board/ committee member" = "Advisory Board Attendee",
    "advisory board/committee member overseas" = "Advisory Board Attendee",
    "committee member" = "Advisory Board Attendee",
    "advisory board / committee member - speaker" = "Advisory Board Attendee",
    "advisory board chair" = "Advisory Board Attendee",
    "committee meeting" = "Advisory Board Attendee",
    "advisory board fees" = "Advisory Board Attendee",

    # Company Meeting Attendee
    "company meeting" = "Company Meeting Attendee",
    "company meeting in australia" = "Company Meeting Attendee",
    "company meeting overseas" = "Company Meeting Attendee",
    "compnay meeting in australia" = "Company Meeting Attendee",

    # Independent Meeting Attendee
    "independent meeting" = "Independent Meeting Attendee",
    "independent meeting in australia" = "Independent Meeting Attendee",
    "independent meeting overseas" = "Independent Meeting Attendee",
    "independentmeeting inaustralia" = "Independent Meeting Attendee",
    "independentmeeting overseas" = "Independent Meeting Attendee",
    "independent overseas meeting delivered online" = "Independent Meeting Attendee",
    "independent virtual meeting overseas" = "Independent Meeting Attendee",

    # Educational Meeting Speaker or Chairperson
    "educational meeting speaker" = "Educational Meeting Speaker or Chairperson",
    "educational meeting chairperson" = "Educational Meeting Speaker or Chairperson",
    "education meeting speaker or chair person" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker or chair person overseas" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker/chair" = "Educational Meeting Speaker or Chairperson",
    "chairperson" = "Educational Meeting Speaker or Chairperson",
    "chairman / moderator agreement" = "Educational Meeting Speaker or Chairperson",
    "educational meetng speaker or chariperson" = "Educational Meeting Speaker or Chairperson",
    "education meeting\r\r\n\r\r\nspeaker or chair\r\r\n\r\r\nperson" = "Educational Meeting Speaker or Chairperson",
    "speaker at merck educational meeting" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker or chrairperson" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker or chariperson" = "Educational Meeting Speaker or Chairperson",
    "educational speaker or chair person" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker or chair person" = "Educational Meeting Speaker or Chairperson",
    "educational meeting chair person" = "Educational Meeting Speaker or Chairperson",
    "educational meeting overseas" = "Educational Meeting Speaker or Chairperson",
    "educational meeting" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker or chairperson" = "Educational Meeting Speaker or Chairperson",
    "educational meeting attendee (also attended a company meeting immediately before)" = "Educational Meeting Speaker or Chairperson",
    "educational meeting speaker/chairperson" = "Educational Meeting Speaker or Chairperson",

    # Educational Meeting Attendee
    "educational meeting attendee" = "Educational Meeting Attendee",
    "education meeting attendee" = "Educational Meeting Attendee",
    "medical education attendee" = "Educational Meeting Attendee",
    "attendee" = "Educational Meeting Attendee",
    "virtual meeting attendee" = "Educational Meeting Attendee",
    "workshop attendee" = "Educational Meeting Attendee",

    # Consultant Engagement
    "consultant" = "Consultant Engagement",
    "consultant fees" = "Consultant Engagement",
    "consulting service" = "Consultant Engagement",
    "consulting services" = "Consultant Engagement",
    "consulting agreement" = "Consultant Engagement",
    "consulting" = "Consultant Engagement",
    "consultant- speaker" = "Consultant Engagement",
    "consultant  services" = "Consultant Engagement",
    "consultant  service" = "Consultant Engagement",

    # Sponsorships
    "hcp sponsorship international" = "Sponsorship",
    "hcp sponsorship australia" = "Sponsorship",
    "hcp other fees" = "Sponsorship",

    # Market Research
    "market research participant" = "Market Research Participant",

    # Miscellaneous
    "not specified" = "Not Specified"
  )

  data %>%
    mutate(
      type_of_service_lower = str_to_lower(type_of_service),
      standardised_service = service_mapping[type_of_service_lower]
    )
}
